se3_gnn_minimal/requirements.txt
# Core
numpy==1.26.4
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3

# Deep learning
torch>=2.1
torch-geometric>=2.4
torch-scatter>=2.1
torch-sparse>=0.6

# SE(3) equivariance
e3nn>=0.5

# Protein language model
fair-esm>=2.0

# Structure parsing
biopython>=1.81

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Data I/O
requests>=2.31
